fix integration tests build on windows
pty.Start isn't exported on windows which fails build.

I moved interactive repl test into separate file with !windows build tag
